# docs/automation/openapi.yaml
...
openapi: 3.0.1
info:
  title: Gravitee.io AM - Automation API
  description: "Declarative, GitOps-style management of Access Management resources."
  version: 4.12.0-alpha.5
servers:
- url: /management/automation
security:
- BearerAuth: []
- BasicAuth: []
...
paths:
...
  /organizations/{orgId}/environments/{envId}/domains/{domainKey}/reporters:
    get:
      tags:
      - Reporters
      - Domains
      summary: List a domain's reporters
      description: Returns all reporters managed by the Automation API under the domain.
      operationId: automationListReporters
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: List of reporters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationReporter"
        "403":
          description: Permission denied
        default:
          description: Unexpected error
    put:
      tags:
      - Reporters
      - Domains
      summary: Create or update a reporter
      description: Idempotent create-or-update. Uses the key field in the body to
        identify the reporter within the domain.
      operationId: automationCreateOrUpdateReporter
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationReporter"
        required: true
      responses:
        "200":
          description: The created or updated reporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationReporter"
        "403":
          description: Permission denied
        default:
          description: Unexpected error
  /organizations/{orgId}/environments/{envId}/domains/{domainKey}/reporters/{reporterKey}:
    get:
      tags:
      - Reporters
      - Domains
      summary: Get a reporter
      description: Retrieves a single reporter by its key.
      operationId: automationGetReporter
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      - name: reporterKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The reporter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationReporter"
        "403":
          description: Permission denied
        default:
          description: Unexpected error
    delete:
      tags:
      - Reporters
      - Domains
      summary: Delete a reporter
      operationId: automationDeleteReporter
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      - name: reporterKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Reporter deleted
        "403":
          description: Permission denied
        default:
          description: Unexpected error
components:
  schemas:
...
    AutomationReporter:
      required:
      - key
      type: object
      properties:
        configuration:
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        dataType:
          type: string
          readOnly: true
        enabled:
          type: boolean
        key:
          maxLength: 255
          minLength: 1
          pattern: "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
          type: string
        name:
          maxLength: 255
          minLength: 1
          type: string
        system:
          type: boolean
          description: "whether this is the domain's system reporter (immutable after\
            \ creation; when true, only key is required and the reporter is built\
            \ from domains.reporters.default.* and repository system settings)"
        type:
          type: string
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      scheme: basic