# gravitee-am-management-api/gravitee-am-management-api-automation/gravitee-am-management-api-automation-rest/src/main/java/io/gravitee/am/management/handlers/automation/resource/ReportersResource.java
/**
 * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *         http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package io.gravitee.am.management.handlers.automation.resource;

import io.gravitee.am.management.handlers.automation.mapper.AutomationReporterMapper;
import io.gravitee.am.management.handlers.automation.model.AutomationReporter;
import io.gravitee.am.management.service.DomainService;
import io.gravitee.am.model.Acl;
import io.gravitee.am.model.Domain;
import io.gravitee.am.model.ManagedBy;
import io.gravitee.am.model.Reference;
import io.gravitee.am.model.Reporter;
import io.gravitee.am.model.permissions.Permission;
import io.gravitee.am.service.ReporterService;
import io.gravitee.am.service.exception.DomainNotFoundException;
import io.gravitee.am.service.exception.InvalidParameterException;
import io.gravitee.am.service.model.AutomationNewReporter;
import io.reactivex.rxjava3.core.Single;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.PUT;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.container.AsyncResponse;
import jakarta.ws.rs.container.ResourceContext;
import jakarta.ws.rs.container.Suspended;
import jakarta.ws.rs.core.Context;
import jakarta.ws.rs.core.MediaType;
import org.springframework.beans.factory.annotation.Autowired;

import java.util.Optional;

/**
 * Reporters managed individually under a domain.
 *
 * @author GraviteeSource Team
 */
@Tag(name = "Reporters")
public class ReportersResource extends AbstractAutomationResource {

    @Context
    private ResourceContext resourceContext;

    @Autowired
    private DomainService domainService;

    @Autowired
    private ReporterService reporterService;

    @GET
    @Produces(MediaType.APPLICATION_JSON)
    @Operation(operationId = "automationListReporters", summary = "List a domain's reporters",
            description = "Returns all reporters managed by the Automation API under the domain.")
    @ApiResponse(responseCode = "200", description = "List of reporters",
            content = @Content(mediaType = "application/json",
                    array = @ArraySchema(schema = @Schema(implementation = AutomationReporter.class))))
    public void list(
            @PathParam("orgId") String organizationId,
            @PathParam("envId") String environmentId,
            @PathParam("domainKey") String domainKey,
            @Suspended final AsyncResponse response) {

        final var principal = getAuthenticatedUser();
        checkAnyPermission(principal, organizationId, environmentId, Permission.DOMAIN_REPORTER, Acl.LIST)
                .andThen(resolveDomain(environmentId, domainKey))
                .flatMap(domain -> reporterService.findByReference(Reference.domain(domain.getId()))
                        .filter(reporter -> reporter.isManagedBy(ManagedBy.AUTOMATION_API))
                        .sorted((o1, o2) -> String.CASE_INSENSITIVE_ORDER.compare(
                                nullToEmpty(o1.getAutomationKey()), nullToEmpty(o2.getAutomationKey())))
                        .map(AutomationReporterMapper::toAutomationReporter)
                        .toList())
                .subscribe(response::resume, response::resume);
    }

    @PUT
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
    @Operation(operationId = "automationCreateOrUpdateReporter",
            summary = "Create or update a reporter",
            description = "Idempotent create-or-update. Uses the key field in the body to identify the reporter within the domain.")
    @ApiResponse(responseCode = "200", description = "The created or updated reporter",
            content = @Content(mediaType = "application/json",
                    schema = @Schema(implementation = AutomationReporter.class)))
    public void createOrUpdate(
            @PathParam("orgId") String organizationId,
            @PathParam("envId") String environmentId,
            @PathParam("domainKey") String domainKey,
            @Valid @NotNull AutomationReporter definition,
            @Suspended final AsyncResponse response) {

        final var principal = getAuthenticatedUser();
        final String domainId = AutomationIds.domainId(environmentId, domainKey);
        final String key = definition.getAutomationKey();
        final Reference reference = Reference.domain(domainId);
        reporterService.findByReference(reference).toList().flatMap(allExisting -> {
            Optional<Reporter> match = allExisting.stream()
                    .filter(reporter -> reporter.isManagedBy(ManagedBy.AUTOMATION_API))
                    .filter(reporter -> key.equals(reporter.getAutomationKey()))
                    .findFirst();
            // The ACL is chosen from a non-erroring lookup so the permission gate runs before any
            // existence is revealed (domain 404 / conflict 400).
            Acl requiredAcl = match.isPresent() ? Acl.UPDATE : Acl.CREATE;
            return checkAnyPermission(principal, organizationId, environmentId, Permission.DOMAIN_REPORTER, requiredAcl)
                    .andThen(resolveDomain(environmentId, domainKey))
                    .flatMap(domain -> {
                        if (match.isPresent()) {
                            Reporter existing = match.get();
                            // 'system' is an immutable identity attribute; reject a change.
                            if (definition.isSystem() != existing.isSystem()) {
                                return Single.error(new InvalidParameterException(
                                        "The 'system' flag is immutable for an existing reporter '" + key
                                                + "'; delete and recreate it to change it"));
                            }
                            if (existing.isSystem()) {
                                // re-PUT is an idempotent no-op
                                return Single.just(AutomationReporterMapper.toAutomationReporter(existing));
                            }
                            return reporterService.update(reference, existing.getId(),
                                            AutomationReporterMapper.toUpdateReporter(definition), principal, false)
                                    .map(AutomationReporterMapper::toAutomationReporter);
                        }

                        final String reporterId = AutomationIds.reporterId(domain.getId(), key);
                        Optional<Reporter> occupant = allExisting.stream()
                                .filter(reporter -> reporterId.equals(reporter.getId()))
                                .findFirst();
                        if (occupant.isPresent()) {
                            return Single.error(new InvalidParameterException(
                                    "Reporter key '" + key + "' conflicts with an existing reporter"));
                        }
                        if (definition.isSystem() && allExisting.stream()
                                .anyMatch(reporter -> reporter.isManagedBy(ManagedBy.AUTOMATION_API) && reporter.isSystem())) {
                            return Single.error(new InvalidParameterException(
                                    "The domain already has a system reporter"));
                        }
                        if (definition.isSystem()) {
                            return reporterService.createSystem(reference, reporterId, key, principal)
                                    .map(AutomationReporterMapper::toAutomationReporter);
                        }
                        Single<AutomationReporter> rejection = rejectIfMissingReporterFields(definition, key);
                        if (rejection != null) {
                            return rejection;
                        }
                        AutomationNewReporter newReporter = AutomationReporterMapper.toNewReporter(definition);
                        newReporter.setId(reporterId);
                        return reporterService.create(reference, newReporter, principal, false)
                                .map(AutomationReporterMapper::toAutomationReporter);
                    });
        })
                .subscribe(response::resume, response::resume);
    }

    private static Single<AutomationReporter> rejectIfMissingReporterFields(AutomationReporter definition, String key) {
        if (isBlank(definition.getName())) {
            return Single.error(new InvalidParameterException(
                    "Field 'name' is required for a non-system reporter '" + key + "'"));
        }
        if (isBlank(definition.getType())) {
            return Single.error(new InvalidParameterException(
                    "Field 'type' is required for a non-system reporter '" + key + "'"));
        }
        if (isBlank(definition.getConfiguration())) {
            return Single.error(new InvalidParameterException(
                    "Field 'configuration' is required for a non-system reporter '" + key + "'"));
        }
        return null;
    }

    @Path("/{reporterKey}")
    public ReporterResource getReporterResource() {
        return resourceContext.getResource(ReporterResource.class);
    }

    private Single<Domain> resolveDomain(String environmentId, String domainKey) {
        return domainService.findById(AutomationIds.domainId(environmentId, domainKey))
                .switchIfEmpty(Single.error(() -> new DomainNotFoundException(domainKey)))
                .flatMap(domain -> domain.isManagedBy(ManagedBy.AUTOMATION_API)
                        ? Single.just(domain)
                        : Single.error(new DomainNotFoundException(domainKey)));
    }
}